[1231] Move Workbench related code to @eclipse-sirius/sirius-components-core
Bug: https://github.com/eclipse-sirius/sirius-components/issues/1231
