+++ frontend/src/index.ts
@@ -12,7 +12,6 @@
  *******************************************************************************/
 // Required because Sprotty uses Inversify and both frameworks are written in TypeScript with experimental features.
 import 'reflect-metadata';
-export * from 'materialui';
 export * from './common/dataTransferTypes';
 export * from './common/sendFile';
 export * from './common/ServerContext';
@@ -20,7 +19,6 @@ export * from './common/URL';
 export * from './core/contextmenu/ContextMenu';
 export * from './core/file-upload/FileUpload';
 export * from './core/form/Form';
-export * from './core/panels/Panels';
 export * from './diagram/DiagramWebSocketContainer';
 export * from './diagram/DiagramWebSocketContainer.types';
 export * from './errors/ErrorBoundary';
@@ -33,10 +31,5 @@ export * from './modals/new-root-object/NewRootObjectModal';
 export * from './modals/rename-project/RenameProjectModal';
 export * from './modals/share-diagram/ShareDiagramModal';
 export * from './modals/upload-document/UploadDocumentModal';
-export * from './theme';
+export * from './onboarding/OnboardArea';
 export * from './views/FormContainer';
-export * from './workbench/RepresentationContext';
-export * from './workbench/RepresentationNavigation';
-export * from './workbench/Workbench';
-export * from './workbench/Workbench.types';
-export * from './workbench/WorkbenchViewContribution';
