+++ packages/core/frontend/sirius-components-core/src/materialui.ts
