+++ packages/core/frontend/sirius-components-core/src/theme.ts
