+++ packages/core/frontend/sirius-components-core/src/workbench/RepresentationContext.ts
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021 Obeo.
+ * Copyright (c) 2021, 2022 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -10,11 +10,17 @@
  * Contributors:
  *     Obeo - initial API and implementation
  *******************************************************************************/
-import { Representation } from 'workbench/Workbench.types';
+import React from 'react';
+import { RepresentationComponentRegistry } from './RepresentationContext.types';
+import { Representation, RepresentationComponentProps } from './Workbench.types';
 
-export type RepresentationNavigationProps = {
-  representations: Representation[];
-  displayedRepresentation: Representation;
-  onRepresentationClick: (representation: Representation) => void;
-  onClose: (representation: Representation) => void;
+const registry: RepresentationComponentRegistry = {
+  getComponent: (representation: Representation) => {
+    return (props: RepresentationComponentProps) => null;
+  },
 };
+
+const value = {
+  registry,
+};
+export const RepresentationContext = React.createContext(value);
