+++ packages/core/frontend/sirius-components-core/src/workbench/Site.tsx
